Cashier/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
fastapi==0.104.0
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
protobuf==4.21.12
pydantic==2.4.2
pydantic_core==2.10.1
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.2.0
uvicorn==0.23.2
